+++ Johnson/Cards/RevampCard.cs
@@ -77,13 +77,12 @@ internal sealed class RevampCard : Card, IRegisterable
 
 				if (Choose)
 				{
-					// TODO: disallow cancel
-					c.QueueImmediate(ModEntry.Instance.KokoroApi.TemporaryUpgrades.MakeChooseTemporaryUpgradeAction(card.uuid).AsCardAction);
+					c.QueueImmediate(ModEntry.Instance.KokoroApi.TemporaryUpgrades.MakeChooseTemporaryUpgradeAction(card.uuid).SetAllowCancel(false).AsCardAction);
 				}
 				else
 				{
 					didAnything = true;
-					ModEntry.Instance.KokoroApi.TemporaryUpgrades.SetTemporaryUpgrade(card, upgrade);
+					ModEntry.Instance.KokoroApi.TemporaryUpgrades.SetTemporaryUpgrade(s, card, upgrade);
 				}
 			}
 		}